requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-nested-admin==4.0.2
django-mysql==4.9.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.5.0
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
yarl==1.9.1
